feat(theme): pass correct mdx components to each template
