+++ websites/docs-smoke-test/src/releases/index.mdx
@@ -2,8 +2,11 @@
 title: Docs Kit Release Notes Test
 ---
 
-<Subtitle>Custom introductory content above the Release notes</Subtitle>
+<Subtitle>
 
-To stay informed about the latest releases subscribe to the RSS feed or sign up for the (dummy product newsletter)[https://example.com].
+Custom introductory content above the Release notes
 
-`<ReleaseNotes />` (JSX tag as code until implemented to not break the build until then)
+</Subtitle>
+
+To stay informed about the latest releases subscribe to the RSS feed or sign up for the [dummy product newsletter](https://example.com).
+
